feat: Update project structure and add dotenv configuration
- Added dotenv configuration for environment variables
- Moved notFound middleware to middlewares/ directory
- Organized health-related functionality:
  - Moved healthController.js to controllers/ directory
  - Moved healthyRoute.js to routes/ directory
- Updated import paths to reflect new directory structure
- Updated README.md with latest project structure details
