+++ src/controllers/healthController.js
@@ -1,0 +1,7 @@
+exports.getHealthStatus = (req, res) => {
+  res.status(200).json({
+    health: "live",
+    status: res.statusCode,
+    message: "ok"
+  });
+};
